Rector as a hook
