+++ app/screens/channel/channel.tsx
@@ -4,7 +4,6 @@
 import React, {useCallback, useEffect, useRef, useState} from 'react';
 import {BackHandler, LayoutChangeEvent, NativeEventSubscription, StyleSheet, View} from 'react-native';
 import {KeyboardTrackingViewRef} from 'react-native-keyboard-tracking-view';
-import {Navigation} from 'react-native-navigation';
 import {Edge, SafeAreaView, useSafeAreaInsets} from 'react-native-safe-area-context';
 
 import CurrentCallBar from '@calls/components/current_call_bar';
@@ -17,6 +16,7 @@ import {ACCESSORIES_CONTAINER_NATIVE_ID} from '@constants/post_draft';
 import {useChannelSwitch} from '@hooks/channel_switch';
 import {useAppState, useIsTablet} from '@hooks/device';
 import {useDefaultHeaderHeight} from '@hooks/header';
+import {useKeyboardTrackingPaused} from '@hooks/keyboard_tracking';
 import {useTeamSwitch} from '@hooks/team_switch';
 import {popTopScreen} from '@screens/navigation';
 import EphemeralStore from '@store/ephemeral_store';
@@ -36,6 +36,7 @@ type ChannelProps = {
 };
 
 const edges: Edge[] = ['left', 'right'];
+const trackKeyboardForScreens = [Screens.HOME, Screens.CHANNEL];
 
 const styles = StyleSheet.create({
     flex: {
@@ -60,32 +61,10 @@ const Channel = ({
     const switchingChannels = useChannelSwitch();
     const defaultHeight = useDefaultHeaderHeight();
     const postDraftRef = useRef<KeyboardTrackingViewRef>(null);
-    const isPostDraftPaused = useRef(false);
     const [containerHeight, setContainerHeight] = useState(0);
-
     const shouldRender = !switchingTeam && !switchingChannels && shouldRenderPosts && Boolean(channelId);
 
-    useEffect(() => {
-        const commandListener = Navigation.events().registerCommandListener(() => {
-            if (!isPostDraftPaused.current) {
-                isPostDraftPaused.current = true;
-                postDraftRef.current?.pauseTracking(channelId);
-            }
-        });
-
-        const commandCompletedListener = Navigation.events().registerCommandCompletedListener(() => {
-            const id = NavigationStore.getNavigationTopComponentId();
-            if ([Screens.HOME, Screens.CHANNEL].includes(id) && isPostDraftPaused.current) {
-                isPostDraftPaused.current = false;
-                postDraftRef.current?.resumeTracking(channelId);
-            }
-        });
-
-        return () => {
-            commandListener.remove();
-            commandCompletedListener.remove();
-        };
-    }, [channelId]);
+    useKeyboardTrackingPaused(postDraftRef, channelId, trackKeyboardForScreens);
 
     useEffect(() => {
         let back: NativeEventSubscription|undefined;
